Replace ObservablePlot modal with Echarts
