+++ package.json
@@ -18,9 +18,7 @@
     "@googlemaps/js-api-loader": "^1.16.6",
     "@googlemaps/markerclusterer": "^2.5.3",
     "@mdi/js": "^7.2.96",
-    "@observablehq/plot": "^0.6.10",
     "country-list": "^2.3.0",
-    "d3": "^7.8.5",
     "date-fns": "^2.30.0",
     "dotenv": "^16.0.3",
     "echarts": "^5.5.0",
