+++ vite.config.ts
@@ -52,9 +52,9 @@ export default defineConfig({
         '**/src/services/handle401.ts',
         '**/src/utils/mdi-icons.ts',
         '**/src/utils/materialColors.ts',
-        '**/src/utils/FocusContextPlot.ts',
         '**/src/utils/CSVDownloadUtils.ts',
         '**/src/utils/plotting/echarts.ts',
+        '**/src/utils/plotting/graphSeriesUtils.ts',
         '**/src/utils/test/**',
         '**/src/utils/googleMaps/**',
         '**/src/utils/rules.ts',
